workspace/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
markdown>=3.5.0
requests>=2.31.0
